[FIX] account: speedup l10n install
The parent_path computation was done on the whole table, being the
principal cause of slowness. Moreover, _adapt_parent_account_group was
updating the whole table, changing table statistics and leading to non
optimal query plan with query lasting more than one minute in some cases.

It was also discovered by ryv that the slowness may have been caused by
a change in parent_tore_update, because the logic was adapted to work
event if records are not already grouped by parent_id.
The previous implemntation was actually leading to invalid result in
this case. The parent path in previous version is actually invalid.
We can safely remove it to solve both issues. This is done by removing
parent_store on the class for easy backport, field can be removed in a
followup pr with upgrade script.