[IMP] account: avoid hudge ids list
_adapt_accounts_for_account_groups and _adapt_parent_account_group are
both called with a lot of ids.

_adapt_accounts_for_account_groups was using account_ids to generate a
domain. In load, this was all account, meaning that the generated query
had an id in where the whole table was contained (at least 13000 ids on
the latest modules).

_adapt_parent_account_group was using self. In _load, self was a search
on all record, then mapped on the company id. This can be simplified by
checking all without a crafted domain.

In both case a boolean params was added. Not ideal, but it could be
interressting to remove those call on the whole table by only checking
the record that could have been impacted, if there was. Some of those
check are actually no op and it seems they are there in case anything
was updated if the flag `delay_account_group_sync` is set in the context

A basic computed field would have triggered dependencies and stored
fields would have been invalidated and recomputed uppon flush. Even if
expressing the dependency is not easy, it could be done manually by
triggering the to_compute maybe. To discuss later.