[IMP] account: avoid bid ids list in select
_adapt_accounts_for_account_groups and _adapt_parent_account_group are
both called with a lot of ids.

_adapt_accounts_for_account_groups was using account_ids to generate a
domain. Ids are comming from a search with all account of the company,
meaning that the generated query had an `id in` where a big part of
table was contained (at least 13000 ids on the latest modules).
This is redundant with a part of the domain that filters on the company.

_adapt_parent_account_group was using self. In _load, self was a search
on all record, then mapped on the company id. This can be simplified by
passing the company id by clarity and simplicity