[FIX] web:  will reflect accurte date during popover
Before this commit:
Whenever the user shifted from there local timezone to some other timezone
like `US/Pacific` which is `UTC-7` hrs and then hovered on a particular
gantt pill then that pill reflects inconsistent date and time.(inconsistencies
in date and time is for models `project.project` and `hr.leave` of modules
`project_enterprise` and `hr_appraisal` where the date range fields are of type
`date` and not `datetime`).

Steps to produce:
- Switch your laptops time zone to some other for eg: `US/Pacific`.
- Project > Configuration > Settings > Task Management, Enable `Project Stages`.
- Project > Projects, change the view to `Gantt` View.
- Now on one of the projects (eg- `Office Design`) hover.
- Popover would appear but it would reflect inaccurate time.

After this commit:
If the user changes there timezone  and then hovers on any particular gantt pill
it would reflect the accurate date and time of the pill in the gantt view.

Task-3424435