wip [FIX] utm: widen search of existing utm source names
[Reproduce]
 - Install marketing_automation
 - Have Campaign with a Template
	 - Access the Campaign's Template and go to the Settings tab.
	 - Note down the NAME of the `utm.source` (found in the Advanced section, under the field "Name").
	 - Delete this template (which will also delete associated `mailing.mailing` records.)
 - Create new template (by returning to the Campaign form, opening an activity, creating template)
 - Open the Campaign's Template:
	 - Disable read-only on the Name field in Advanced section (modify the `mailing.mailing.form` view).
	 - Change its name to the previously noted `utm.source` NAME.
	 - Save -> BUG: Trackback "The name must be unique"

[Why this bug happens?]
When you delete a template (`mailing.mailing`) related to a `utm.source`, the `utm.source` will still exist.
Then, when attempting to change the `utm.source` name, the system is supposed to detect if the name already
exists and generate a new unique name if necessary. However, due to the mechanism utilized in the
function (ref.1), which searches for existing names based on existing templates (`mailing.mailing`),
which you have deleted, the existing `utm.source` NAME won't be found. Consequently, the function mistakenly
believes the name is unique and doesn't generate a new name, leading to the encountered error.

(ref.1) `UtmMixin._get_unique_names`
https://github.com/odoo/odoo/blob/205caa3d6539187009a8b8bbd1557e75fe045ff3/addons/utm/models/utm_mixin.py#L77

opw-3835704