[FIX] mail: keep active discuss thread after reload
Before this commit, reloading the page will loose the current
thread of the discuss app. This issue is a side effect of the
recent URL reformatting ([1]). Discuss was relying on the active_id
string to retrive the thread model and id after a page reload.

Since [1], the path name of the URL is interpreted to deduce
several information from it: action, active_id, resId and
model. As a result, the active_id must be an integer and any
string containing a dot is interpreted as the model part of the
url.

To sum it up:
- active_id cannot be used as we need a combo model/id to retrieve
the thread.
- model cannot be used as it is ignored when an action is present.

As a result, the only remaining option is a query string parameter.

task-3853219