[IMP] web:  will reflect the same page even if we switch views
Before this commit:
If the user was in a particular view let us suppose (Kanban view). After
enabling the pager when the user shifts to next page and then changes the view
to list view from kanban view then the data is loaded from the beginning.

After this commit:
If the user switches to another view then then also the pager values will remain
same and user will be on same page in another view.

Task-3845097