[ADD] l10n_it_edi_doi: Declaration of Intent support for l10n_it
In Italy, an operator who exports more than 10% of all his revenues is considered a regular exporter.
Regular exporters can purchase without paying VAT within certain limits.

For this they have to do a declaration of intent with the authorities.
The regular exporter then gives the declaration information to his suppliers.
The supplier can then issue an invoice to the regular exported
without taxes by noting the declaration of intent.

This commit adds support for this use case.
It is now possible to specificy a declaration of intent on sales orders and invoices.
They are also stated on the PDF and electronic invoices.
A dedicated tax and fiscal position was added to be applied in case a declaration of intent is used.

Limitations:
- The document currency has to be in EUR
  (more precisely the currency of the declaration of intent which defaults to EUR);
  else the declaration of intent cannot be applied
- Lines containing the declaration of intent tax may not use any other tax
- When adding lines on a credit note linked to a sales order those lines do not count
  for the invoiced amount of the sales order. The sales order would have to be adapted.
  (It is not really a limitation because that is just how it works: the new lines are not linked to the sales order.)
- Documents can only use the declaration of intent if they lie inside the date range of the declaration of intent
  (This may be a problem if one wants to create refunds at a later time)
- On (confirmed) sales orders the warning amount is only updated after saving (technical limitation)
page.

task-3794632