[FIX] point_of_sale: runbot: check order is selected
During test, the clicked order may not be properly selected when the numpad is
pressed for refund. This change makes sure that we only proceed on clicking the
numpad when the selected order reflected in the screen.

Failing tour: FiscalPositionNoTaxRefund
Failing test: test_refund_order_with_fp_tax_included
Runbot: 61512