[FIX] point_of_sale: number buffer when pressing +/-
Before this commit:
When we clicked on the +/- button on the numpad, the number buffer wasn't
modified correctly. As a result, we couldn't obtain the necessary changes,
and the field value on the screen would revert to 0.

After this commit:
Now on pressing the +/- button now yields the correct value, and properly
handled the buffer value.

task- 3869556