[FIX] website_sale_loyalty: prevent error while remove product from cart
Currently, error is generated when remove a reward product that has been added
multiple times  in cart

Step to produced:

- Install a 'website_sale_loyalty' module.
- Navigate to the website / eCommerce / Loyalty / Discount & Loyalty create
a record.
- Set the Loyalty Program name and Program Type as 'Loyalty Cards'.(Ensure it's
available on sale and on the website.)
-  And add 'Rewards' and set a Reward Type set as 'Discount' which is  apply to
on Cheapest Product.
 - Go to website shop add a any product on card, Open a cart increase the
 quantity of the product and claim the discount reward multiple times.
- Now try to remove a reward product that has been added multiple times  in
cart.

The issue occurs when attempting to 'Remove' reward product which is added
multiple time in cart. At this point [1], system try to get 'id' from multiple
lines.

To resolve the issue, Add indexing  so system will now target and remove only
the one instance of the duplicate product from the cart.

sentry-5091141573