[FIX] fields: fix traceback when user gives string as default value for ID in studio
This traceback occurs when the user tries to give a string as the default value
through studio.

To reproduce this issue:

1) Install studio and inventory module
2) Open any product in studio mode
3) Activate 'show invisible elements' from views in the studio sidebar
4) Now click on id in from viewand set the default value as a string
5) A traceback occurs

Error:- "TypeError: '<' not supported between instances of 'int' and 'str'"

Because when the user gives a string as default value it tries to parse
the field with the 'convert_to_cache' method.

But in our case ID class does not have the method 'convert_to_cache'
so it gives the parse value as it is.

This leads to the above traceback when the default value (string) is tries to
compare with integers.

https://github.com/odoo/odoo/blob/871121ab4a7a4ed5fc309823c5c41ccfeff7536d/odoo/addons/base/models/ir_default.py#L73-L82

sentry-5193548638