[IMP] base: support all encryptions in --smtp-ssl
The `--smtp-ssl` was used as a True/False flag, upon activation it was
selecting the "starttls" smtp encryption method. With the recent new
"starttls_strict" and "ssl_strict" encryption methods, it deemed
important to let users select their prefered encryption method via the
CLI.

The option `--smtp-ssl` can now be used either as a flag (in which case
it selects "starttls") but it can also be used as a regular option where
the desired encryption method follows the option name. The supported
options are the same as the `smtp_encryption` selection field of
`ir.mail_server`, they are: `none`, `ssl`, `ssl_strict`, `starttls`, and
`starttls_strict`. Note that `--smtp-ssl none` is the same as no option.

Note: because of a limitation of optparse, it is not possible to supply
the option using `--smtp-ssl=value`, the option and its argument must be
separated by a single space: `--smtp-ssl value`.

task-2861790