[IMP] website_forum: improve forum promenade
This commit makes some slight visual changes to the forum pages, to
improve their readability:

forum index page:
----------------

In the sidebar, the title of the 'Posts' link is renamed to 'All Posts'
and the 'View all' link is aligned on the left, close to the 'Tags'
title. Links to posts already visited are lightened by 50%, making them
difficult to read. They are now lightened to 25%.

tags page:
----------

In this page, the user can choose to follow or stop following a
particular tag. Followed and unfollowed tags have a different style,
allowing the user to distinguish between them. But the style is updated
on the server side. This commit ensures that this visual update takes
place on the front-end following the user's action.

post creation page:
------------------

2 additional tips are added at the bottom of the description block: "Use
'/' to insert images, Select text to format it". Also, the option to
create a new tag is displayed last in the tag selection list, to limit
the creation of duplicate tags.

post consultation page:
----------------------

When the user accesses the page, the background of the question block is
highlighted in light blue for 2 seconds. The vertical/horizontal spacing
of certain elements has been reduced: between the post title and the tag
list, the favorite icon and the follow icon... View and reply counters
have been shifted to the right. The opacity of voting buttons has been
increased. The tooltip for the 'Edit your answer' button has been
removed, and the option of converting a reply into a comment has been
added inside the dropdown for additional actions.

Finally, addition of a primary color outline on the 'Answer' button at
the bottom of the page. The 'fa-reply' icon was replaced, as in this
context the 'Answer' button looked more like a link back to the previous
page.

task-3859111