[IMP] website_forum: add a related post section
Purpose:
=======

This commit adds a list of 5 posts at the bottom of the post page.
This feature aims at providing users with additional content to read
similar to the post they are consulting.

Specs:
======

The feature relies on the '_prepare_related_posts' called by the
'website_forum' controller when the post page is requested.
The method returns a list of related posts based on tag similarity.

To help test this new feature, a test has beed added and the forum
and forum post populates have been updated with tag creation.

task-3859111