[IMP] web, *: generate PWA manifest on demand
*: hr_attendance, point_of_sale, pos_self_order

This commit implements a mechanism to install Progressive Web Apps in more places
in Odoo. A common route can be used to generate a webmanifest on demand, allowing
reusability and a more common behavior of PWAs accross Odoo.

For backend apps, a dedicated page with a different URL (without the /odoo scope) must
be used. This page has a UI explaining the app, its features and an Install button that
launches the native install dialog. On Safari, instead of the Install button, since
the install dialog does not exist, we explain the steps to add the app to the Home screen
or Dock.

By using this common route, we can make sure the manifest has the right attributes and
we offer the same experience for most use cases (app icon, dedicated scope, engaging
install UI, etc.).

In this commit, we add a support for a few applications. But the feature could easily be
expanded to support more apps and scenarios.

It is possible to give some unique dedicated features to apps, such as app shortcuts.