[FIX] FigureComponent: only propagate CTRL+Y/Z
Previously, when a figure was selected, And pressing CTRL + A would
select all rows and columns in the sheet, and pressing the Delete key
would delete the selected figure. This occurred because all key events
were bubbling up to the grid component.

This commit now fixes the problem by only passing key events
for CTRL + Y/Z to the grid element.

Task: 3863300