[FIX] website: fix default height on image gallery
Steps to reproduce:

- Go to website > Drop an "Image Gallery" snippet > Add images > The
snippet has a default height value displayed in the "Height input" (70%
of the screen height).
- Change the height value of the snippet > Select an image > change its
order in the carousel > The height of the snippet is automatically
re-updated (the height is also overridden in the "Height input").

A 14.0 fix was made in [1] to be sure that "Image Gallery" snippet
starts with a forced height of 70% of the current screen height then the
user will have the possibility to edit the height value using the input
in the panel... But since it's the `slideshow()` method that does the
height update, the value will always be overridden to 70% of
`window.innerHeight` every time we update the snippet layout (e.g. after
re-order).

The goal of this commit is to maintain height consistency by simply
removing the code that updates the height in JS. This way, the snippet
will start with a hardcoded default height value (500px) that is
updatable using the "Height input".

Remark: a possible solution is to move the JS height update to the
`onBuilt()` method, but this will only work for 14.0 since starting from
15.0, the snippet is already rendered with default content and as a
result, the `onBuilt()` height update will lead to a "flicker" effect...

Spotted while reviewing [2].

[1]: https://github.com/odoo/odoo/commit/d5d138e833344e857a420d865d4b12f1acdb0e7c
[2]: https://github.com/odoo/odoo/pull/117369

task-3414281