[FIX] base: prevent copying key data on duplication
When a user duplicates a view of the website menu and subsequently navigates
to the website and click on edit the traceback will appear.

Steps to reproduce:
- Install the 'website' module
- Website > Site > Pages
- Tick any one page
- Actions > Duplicate
- Repeat 2-3 times
- Click on the selected page or a duplicated one
- Edit

Traceback: -
Expected singleton: ir.ui.view(2224, 2225)

This commit will fix the above error by preventing the copy of the associated
data with the key.

sentry - 4991676595