[IMP] web: update action service context evaluation order
This commit will inverse the context evaluation context order of the
action service so that the additional context passed to the action service
overwrites the action context. This change should enable the user to easily
enable/disable filters, update the active ids stored in the context, etc.
without having to modify the action in-place or creating a new action.