[FIX] pos_loyalty: compute points with subset of order lines
Currently when applying a coupon for 10% on the order on the current order, the points awarded are 100% of the points instead of 90%. The loyalty program rewarding x points per $ spent.

Steps to reproduce:
-------------------
* Go to the **Point of Sale** App
* Open shop session
* Select a customer
* Add a product (ex. Office chair 70$)
* Select **Enter Code**
* Enter the code `10pc`
> Observation: The order will grant 700 points instead of 630. The amount paid is 63$ with the code for 10% on the order.

Settings of the **Discount & Loyalties**:
* **Code for 10% on orders**
  * Rule:
    * `Among` -> no restriction set
    * `Minimum quantity`: 1
    * `Minimum Purchase`: 0
  * Reward:
    * `Discount`: 10% `on` order
* **Loyalty Program**
  * `Use points on`: Current & Future orders
  * Rule:
    * `Among` -> no restriction set
    * `Minimum quantity`: 1
    * `Minimum Purchase`: 0
    * `Grant` 10 loyalty poins per $ spent

Why the fix:
------------
The general loop to compute points goes as follows:

```js
for (const program of programs) {
  for (const rule of program.rules) {
    for (const line of orderLines) {
      ...
    }
  }
}
```

For more details check: https://github.com/odoo/odoo/blob/3fd308981d077bd9b93394d45defaa990d9c7acd/addons/pos_loyalty/static/src/js/Loyalty.js#L886

https://github.com/odoo/odoo/blob/3fd308981d077bd9b93394d45defaa990d9c7acd/addons/pos_loyalty/static/src/js/Loyalty.js#L921-L924

Currently this condition is what causes the discount to not count in the computation of the points as the order line associated with the 10% discount coupon is a reward line. Simply removing `orderedProductPaid += line.get_price_with_tax();` out of the if statement is not an option. Doing this causes unwanted behavior in the computation of loyalty points if the points are used and the reward is x $ discount per point used.

https://github.com/odoo/odoo/blob/3fd308981d077bd9b93394d45defaa990d9c7acd/addons/pos_loyalty/static/src/js/Loyalty.js#L908-L914

This is never true as `reward.program_id` should actually be `reward.program_id.id`. Fixing this would also result in the behaior noticed as the reward (the 10% discount) is not part of the loyalty program.

Those two piece of code could be changed to
```js
if (line.is_reward_line) {
  const reward = this.pos.reward_by_id[line.reward_id];
  if (program.id === reward.program_id.id) {
      continue;
  }
}
```

and

```js
orderedProductPaid += line.get_price_with_tax();
totalProductQty += lineQty;
```

without the if statement.

With this we never give points for reward line awarded by the current program but we still count discounts for the computation of points awarded.

opw-3858286