[IMP] base: add batching fetch method
As per tickets 3859206,3332771,3349879, there is a need to batch a dictfetchall
call if the number of rows x columns generated by a query is very large.

This batching method uses a generator function that makes use of `dictfetchmany`
and prevents overutilzation of memory by limiting the allocations made in `__build_dict`.