[IMP] web: Convert ESLint configuration from JSON to JS file format
This commit changes the file format used to configure ESLint from JSON
to JS.

The goal of this change is to allow the addition of custom plugins and
rules to our configuration. Custom plugins are defined in JavaScript,
but ESLint doesn't allow you to reference a JS configuration file from
within a JSON configuration file.

Note: We decided not to use the new ESLint "flat format" because the
integration with VSCode is still not great. Namely, a flag needs to be
explicitly enabled for VSCode to look for eslint.config.js (the new flat
format filename), and even with the flag enabled, it won't find it
unless the file is directly in the directory open in VSCode; which is a
problem, since you usually open the parent directory to have both
Community and Enterprise repositories.

Part of task-3869533.