add command.isDisabled/isSelectionInBlockRoot and use them where missing
The Powerbox commands have a parameter to disable them in certain
contexts, and many had it set on `!isSelectionInBlockRoot()` since
commit [1] but since neither of these things were converted for Phoenix,
many commands were missing the `isDisabled` key.
This applies them in all the places that were incompletely converted but
we'll need to make sure we don't forget to add them to future
conversions.

[1]: https://github.com/odoo/odoo/commit/defce6b8339862c0efd355823142bad3ca81f90a