[IMP] t9n: add custom views to the existing models.
This commit adds the full flow starting from the home screen. The home
screen contains a list of all projects that you can click on to take you
to the list of available target languages in this project. Clicking on a
language will take you to the resources in the project then to the
different texts that need to be translated. Also, various features are
added in each custom view of those like sorting and searching. Finally,
you can add your translation for any text you want :))

Task-3783071