[FIX] discuss: public channel non joined user chat window error
Before this commit:

When users who have not joined the public channel tried to access it through
mention in chatter it threw an error and had new message as channel name instead
of the actual name.

After this commit:

When users who have not joined the public channel try to access it through
mention in chatter it opens without any error and has the
right name of the channel.