[IMP] website_sale_mondialrelay: convert jQuery code to Vanilla JS
This PR aim to convert all jQuery code into Vanilla JS in
website_sale_mondialrelay, this way we will reduce the dependency of jQuery
in Odoo codebase.

This commit also aims to load Mondial Relay Plugin in iframe to remove
dependency of jQuery, the Mondial Relay plugin is built using jQuery,
so we need jQuery to utilize the Mondial Relay API. To achieve this,
we are creating an iframe and loading jQuery within it, along with
explicitly loading the Mondial Relay API. This setup enables us to use
the Mondial Relay delivery method.

task-3770362