[IMP] account: allow change of journal for posted before
Once a move or journal entry is posted, even if it is set back to draft, the journal
cannot be edited. This commit gives the user the option to do so, by allowing the
journal to be changed to a journal of the same type if the name of the move is removed.

This can introduce gaps in the sequence, but that is already allowed behavior when the
user deletes the move name.

task-3829848