[IMP] account: add late payment charges facility
Before this commit:
The facility to charge a penalty on late payments did not exist.

After this commit:
A late payment charge option is provided in payment terms. Penalties can be
applied to both the total or taxable amounts. Six different models are provided
for calculating penalties:

1) Monthly     : Penalty is calculated at the given rate for each month from due
                 date to payment date.
2) Quarterly   : Penalty is calculated at the given rate for each fiscal quarter
                 from due date to payment date.
3) Half-Yearly : Penalty is calculated at the given rate for each fiscal
                 half-year from due date to payment date.
4) Yearly      : Penalty is calculated at the given rate for each fiscal year
                 from due date to payment date.
5) Fixed       : A fixed penalty is charged at the given rate regardless of the
                 payment date.
6) Flat        : A flat amount is charged as a penalty regardless of the payment
                 date.

Late charges are applied if an invoice/bill is overdue. A new debit note is
created with a late payment charges product and its corresponding penalty amount
in draft state, referring to its parent invoice/bill. The late payment charges
product is created when a late penalty is applied for the first time in a
database.

task-3774599