[IMP] pos_restaurant: Instant drag and drop tables
Allow the user to easily manage table plan by moving table in one click.
No more first click to select the table and then another click to be
able to move it.

Instructions : 

Click on edit plan
Click on a table and maintain it
Being able to move the table while the click is maintained

This is valid in desktop. In mobile, the user will still have to click
on the table to select it and then click on it again to move it.