[FIX] stock_picking_batch: adapt batch action done for new quantity
With the quantitypocalypse, the meaning of the 'quantity_done' and
'reserved_availability' has been squashed into a single 'quantity'
field.
The batch `action_done` was not properly updated to match this change,
as it kept ealier behavior distinction between immediate transfers &
partial validation, which are irrelevant now, as the immediate transfers
wizard is no longer a thing.
Also, the 'picked' field on the moves used to separate quantity
reserved/done is readonly on the batch form, so we can't really use it
as distinction in this case.

So updates the picking sanity_check so it could consider all pickings as
one when dealing with batches for the empty picking checks, as it
triggered wrong errors, forbidding to validate batch that was partially
done.

opw-3884043