[FIX] manufacturing: 3 step MRP - Cancel next move not effective
Issue:
==============================
The "Cancel Next Move" option in the production rule is ineffective. Regardless
of its setting, the post-production picking is always cancelled when a
Manufacturing Order (MO) is cancelled.

Resolution:
==================================
To address this issue, the next picking will only be cancelled if all the moves
in the current picking have propagate_cancel set to true when the MO is
cancelled.

Steps to Reproduce:
===============================
1. Configure the 3-step rules for MRP.
2. Check the first rule and untick the "propagate cancel" option.
3. Create a MO, cancel it, and then go to the transfer section. All pickings
   are cancelled regardless.

Expected Result:
================================
After implementing the solution, the next picking will only be cancelled if the
"propagate cancel" option is true for the first rule. Otherwise, next
pickings will not be cancelled when the MO is cancelled.

task-3798448