[FIX] web: display correct week days in month view calendar
Versions
--------
- saas-17.2+

Steps
-----
1. Change browser time zone to America/Los_Angeles;
2. go to Calendar;
3. switch to Month view.

Issue
-----
The day names displayed are one day off, e.g. showing 2024-05-01 as TUE
instead of WED.

Cause
-----
Commit 90f85a19deae updated FullCalendar to V6.1.10.

One of the changes is that month view renders don't look at a specific
date, but one in the distant past, see:
https://github.com/fullcalendar/fullcalendar/issues/5854#issuecomment-702495211

The time zone of this "distant past" should get interpreted as UTC
instead of defaulting to the local time zone.

Solution
--------
In the custom render function passed to FullCalendar, intepret the Date
object as UTC time instead of local time when scale is set to month.

opw-3934014