[FIX] web: werkzeug3.iri_to_uri only takes str
`iri_to_uri` is a function that takes a humain URL (with accents) and
encode it to a wire-compatible URL (with accents encoded). The function
used to be compatible with both bytes and str, but bytes support was
deprecated in werkzeug 2.4 and removed in werkzeug 3.0[^1].

Using the `iri_to_uri` function was only a trick to support both Py2 and
Py3 at the same time. Because Odoo 16 is only compatible with Py3, we
can always use `decode()` to convert the bytes in a str.

[^1]: https://werkzeug.palletsprojects.com/en/3.0.x/urls/