[IMP] web, spreadsheet: Graph view redesign
--- Overall improvements ---
This commit enhances the UI of the Graph view to improve its usability.
This improves the overall design (colors, spacing, etc.) and adapts
colors for darkmode.

--- Prevent chart animation spam ---
Prior to this commit, when a type of chart was already loaded (eg. Pie
chart), you could click on the button of the same type (eg. Pie) to
reload the animation.
This commit prevents the clickling on the active button, when a type of
chart is already loaded.

--- Hide tooltips on mobile ---
To make graphs easier to read in mobile, this commit hides tooltips on
small screens.

task-3620659