[FIX] bus: websocket rate limiting wrongly handling burst
To limit websocket rate, two parameters are available `RL_DELAY` and
`RL_BURST`. The latest specify how many requests can be made in excess
of the given rate at the begining.

Technically, this is implemented using a deque that have the `maxlen`
set to the same value than `RL_BURST`. Once the dequeue is full, we
check if the time between the first frame received and now is
satisfying.

This does not correspond to the description of the burst as the check
is done when there is *exactly* `RL_BURST` frame received. It should
instead be done when the deque is bigger than the accepted burst.