[PERF] account, l10n_*: Improve batch send and print performances
Problem
---------
When doing the Send & Print on one invoice, or when multiple invoices
are selected and the "Download" option is ticked, the PDF generation of
the invoices is made synchronously. Generating a simple invoice takes
~2 seconds.

When you're doing the mass Send&Print for 10+ invoices, you can have
your screen loading for >20seconds just to get 10+pages of PDF.

Objective
---------
Start digging into that part to see if optimisation can be done to
reduce that loading time for UX improvements.

Solution
---------
ONGOING

Results
---------
Before
Nb invoices | time (in sec)
1 | 3.3
5 | 13.9
10| 27.5
20| 55.16

After
Nb invoices | time (in sec)
1 | 3.3
5 | 4.4
10| 5.4
20| 8.4

task-3695628