[IMP] *_livechat: improve and generalize chatbot UTM trackers
RATIONALE

The way we currently use the UTM trackers (aka source/medium/campaign) does not
comply with the industry standards, leading to confusion and globally unclear
or not usable reporting on UTM data.

PURPOSE

Use the recently introduced utm_reference logic to replace the current utm
source when linking a specific originating chatbot.script to leads/tickets.

Also specify and generalize the UTM medium and source to complement reporting.

Now, when creating a crm.lead/helpdesk ticket, the logic is as follows:
- UTM Campaign: Not specified
- UTM Medium: utm.utm_medium_website
- UTM Source: utm.utm_source_chatbot
- UTM Reference: chatbot.script,42

Task-3330036