[FIX] {purchase_}stock: prevent reordering rules for consumable products
Bug:
It was possible to create a reordering rule for a non-storable product
via the import records tool. This is because the domain on the product
(limiting the selection to only storable products) for the reordering
rule model was only enforced on the web client.

Fix:
Enforce this programatically in the `create` method of the model.

Task-3866078