[IMP] history_plugin: execute a resource for plugins to act on a the mutations
Plugins or external components may want to act on the mutations when they occur.
For example, when the editable contains many view parts (the dom is the result of the rendering of some qweb views)
it is useful to intercept the mutations an mark them or their relevant parent as dirty.