[IMP] web, web_editor: emoji picker inherit backend design
This commit aims to prevent the `emoji picker` from inheriting website
theme customizations.

Prior to this commit, the emoji picker was using Bootstrap classes. The
issue with that is that many SCSS variables are overridden
in the front-end, making the style totally different from the expected
result.

To handle this undesired behavior, this commit introduces CSS variables
that will be use to set the desired colors for the front-end.

task-3679384