[FIX] web, web_editor, website, *: handle webp to png implicit conversion *: website_sale
- store the correct mimetype for processed attachments
- suggest jpeg formats when webp is unavailable in snippet options

When processing images to apply modifications with
`HTMLCanvasElement.toDataURL`, safari and some other webkit browsers
return a png when asked for a webp.

Steps to reproduce (4 cases):
- Use safari or an affected browser
- One of:
  - Image Field
    - In any regular page such as the Discuss home page, click on your
      user profile picture (top-right)
    - Click on Preferences
    - Change the profile picture to a webp image
  - Website editor
    - Insert a snippet containing an image
    - One of:
      - Change the image format to one of the webp formats with a lower
        resolution
      - Crop the image
      - Apply a shape mask
    - Save

Old behavior: Processed images are saved with an image/webp mimetype
even though the actual data type is image/png
New behavior: Processed images are saved with the corresponding
mimetype of the actual data

To fix this, handle all possible implicit conversions through a
`convertCanvasToDataURL` and return both the dataURL and the actual
mimetype in `convertCanvasToDataURL` (new util in web
`image_processing.js`) and `applyModifications` (from web_editor
`image_processing.js`). Each case can then deal with mimetype mismatch
in an appropriate way.

Suggest jpeg formats instead of webp in the image snippet options when
the browser is incapable of generating webp with
`HTMLCanvasElement.toDataURL`.

The tests cover the 4 use cases of `applyModifications` and the
standalone use of `convertCanvasToDataURL`.

task-3847470