[FIX] Revert tests (P in lists)
This commit reverts test additions and modifications introduced by FGE's
commit that were originally introduced by [1].

The reason for this is that the spec is about to change, see [2].

So, instead of implementing a behaviour that is about to change, I
chose to not change the normalization of paragraphs inside list items
(and revert the concerned tests so that we can have a green build).

The new behavior shall be applied to phoenix after it gets merged on
stable (PR: [3]).

[1]: https://github.com/odoo/odoo/pull/165470/commits/24be8567b8c038de6f54786e163bc9b8f9f99f96
[2]: https://www.odoo.com/odoo/project/1695/tasks/3932057?cids=1
[3]: https://github.com/odoo/odoo/pull/166433