[IMP] mrp:  Automatic Valuation Company Specific
Before this Commit :
===============
if we select the automatic accounting in the settings,
Inventory valuation field was visible at the product categories,
which is a selection field having two values manual and
automated, but this feature is confusing because no company
selects some categories for automatic and some for manual valuation.

After this Commit :
==============
Inventory valuation field is removed from the product categories form.
and now if  automatic accounting is selected in settings,
then all the product categories will be considered for automatic accounting ,
and if this automatic accounting boolean is unchecked by the user
then it will consider all the product categories as manual.

task - 3973789